fix(workspace): resolve gating issues from PR review
- Extract SIGKILL test functions into process_executor_stream_unix_test.go
  with //go:build !windows constraint (fixes Windows compilation failure)
- Remove syscall import from process_executor_stream_test.go
- Replace WNOHANG tight-loop with blocking Wait4 + throttled fallback
  with time.Sleep(1ms) (fixes CPU-pegging busy-loop)
